feat: 데일리로그 인풋 기능 70% 완료
